Safari-611.1.10/LayoutTests/editing/pasteboard/insert-text-decoration-expected.txt

This tests inserting elements with CSS3 text decoration properties

Content to insert:
| "\n            "
| <p>
|   "Longhands: "
|   <span>
|     style="-webkit-text-decoration-line: overline; -webkit-text-decoration-style: dotted; -webkit-text-decoration-color: blue;"
|     "overline dotted blue"
| "\n            "
| <p>
|   "Shorthand: "
|   <span>
|     style="-webkit-text-decoration: underline dashed red;"
|     "underline dashed red"
| "\n        "

After insertion:
| <p>
|   "Longhands: "
|   <span>
|     style="-webkit-text-decoration-line: overline; -webkit-text-decoration-style: dotted; -webkit-text-decoration-color: blue;"
|     "overline dotted blue"
| "\n            "
| <p>
|   "Shorthand: "
|   <span>
|     style="-webkit-text-decoration: underline dashed red;"
|     "underline dashed red<#selection-caret>"
